ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
joblib==1.3.2
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.5.3
numpy>=1.24.0
pandas>=2.0.0
tensorflow>=2.13.0
yfinance>=0.2.28
matplotlib>=3.7.0
